chore: Move HTTP clients to net folder
